+++ src/main.c
@@ -202,6 +202,8 @@ LRESULT CALLBACK MainWndProc(HWND hwnd, UINT msg, WPARAM wp, LPARAM lp)
     switch (msg)
     {
     case WM_KEYDOWN:
+        /* F1 opens help */
+        if (wp == VK_F1) { Help_Show(); return 0; }
         /* Ctrl+Tab / Ctrl+Shift+Tab to switch tabs */
         if (GetKeyState(VK_CONTROL) & 0x8000) {
             if (wp == VK_TAB) {
@@ -541,6 +543,10 @@ apply_theme:
     case IDM_GITHUB_SCRIPTS:
         ShellExecute(NULL, L"open",
             L"https://github.com/KaiUR/Pycatia_Scripts", NULL, NULL, SW_SHOW);
+        break;
+
+    case IDM_HELP_CONTENTS:
+        Help_Show();
         break;
 
     case IDM_ABOUT: